requirements.txt
aniso8601==7.0.0
asgiref==3.4.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
channels==3.0.4
chardet==4.0.0
charset-normalizer==2.0.6
cloudinary==1.26.0
constantly==15.1.0
cryptography==3.4.8
daphne==3.0.2
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.7
django-cloudinary-storage==0.3.0
django-cors-headers==3.8.0
django-filter==2.4.0
django-graphql-jwt==0.3.1
django-graphql-social-auth==0.1.4
google-api-core==2.0.1
google-api-python-client==2.21.0
google-auth==2.1.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
graphene==2.1.9
graphene-django==2.15.0
graphene-file-upload==1.3.0
graphql-core==2.3.2
graphql-relay==2.0.1
graphql-ws==0.4.4
gunicorn==20.1.0
httplib2==0.19.1
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
oauthlib==3.1.1
Pillow==8.3.2
promise==2.3
protobuf==3.18.0
psycopg2-binary==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.1.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-decouple==3.4
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
Rx==1.6.1
service-identity==21.1.0
singledispatch==3.7.0
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
text-unidecode==1.3
Twisted==21.7.0
txaio==21.2.1
typing-extensions==3.10.0.2
uritemplate==3.0.1
urllib3==1.26.6
zope.interface==5.4.0
